+++ Sources/AENetwork/Helpers/URLRequest+Helpers.swift
@@ -1,7 +1,7 @@
 /**
  *  https://github.com/tadija/AENetwork
- *  Copyright (c) Marko Tadić 2017-2019
- *  Licensed under the MIT license. See LICENSE file.
+ *  Copyright © 2017-2020 Marko Tadić
+ *  Licensed under the MIT license
  */
 
 import Foundation
@@ -101,7 +101,7 @@ public extension URLRequest {
     var fullDescription: String {
         let headers = "\(allHTTPHeaderFields ?? [:])"
         let parameters = "\(url?.parameters ?? [:])"
-        let body = (try? httpBody?.toDictionary()) ?? [:]
+        let body = (try? httpBody?.jsonDictionary()) ?? [:]
         return """
         - Request: \(shortDescription)
         - Headers: \(headers)