+++ Tests/AENetworkTests/URLRequestTests.swift
@@ -1,7 +1,7 @@
 /**
  *  https://github.com/tadija/AENetwork
- *  Copyright (c) Marko Tadić 2017-2019
- *  Licensed under the MIT license. See LICENSE file.
+ *  Copyright © 2017-2020 Marko Tadić
+ *  Licensed under the MIT license
  */
 
 import XCTest
@@ -72,7 +72,7 @@ class URLRequestTests: XCTestCase {
 
         let requestHeaders = "\(request.allHTTPHeaderFields ?? [:])"
         let requestParameters = "\(request.url?.parameters ?? [:])"
-        let requestBody = (try? request.httpBody?.toDictionary()) ?? [:]
+        let requestBody = (try? request.httpBody?.jsonDictionary()) ?? [:]
         let fullDescription = """
         - Request: \(shortDescription)
         - Headers: \(requestHeaders)